requirements.txt
catboost==1.2.8
joblib==1.5.1
lightgbm==4.6.0
matplotlib==3.10.3
numpy==2.3.1
optuna==4.4.0
pandas==2.3.0
Pillow==11.3.0
scikit_learn==1.7.0
seaborn==0.13.2
streamlit==1.46.1
